visualization in new structure (with ion-fab-button routing)
